+++ X-Science/CHANGES.md
@@ -1,6 +1,15 @@
 [x] Science!
 ============
 
+
+4.6
+---
+24-June-2015 Gui bug for KSP V1.0.4
+* Completed experiments total is updated correctly when vehicles are recovered
+* Window is closed on main menu and certain other scenes
+* Added close button to window
+
+
 4.5
 ---
 12-June-2015 Speed increase and bug fix for KSP V1.0.2